refactor: move `CssLoaderError` into own file
