+++ lib/processCss.js
@@ -2,9 +2,9 @@
 	MIT License http://www.opensource.org/licenses/mit-license.php
 	Author Tobias Koppers @sokra
 */
-var formatCodeFrame = require("babel-code-frame");
 var postcss = require("postcss");
 var plugin = require("./plugin");
+var CssLoaderError = require("./CssLoaderError");
 
 module.exports = function processCss(inputSource, inputMap, options, callback) {
   var query = options.query;
@@ -41,43 +41,19 @@ module.exports = function processCss(inputSource, inputMap, options, callback) {
         urlItemRegExp: /___CSS_LOADER_URL___([0-9]+)___/
       });
     })
-    .catch(function(err) {
-      if (err.name === "CssSyntaxError") {
-        var wrappedError = new CSSLoaderError(
-          "Syntax Error",
-          err.reason,
-          err.line != null && err.column != null
-            ? { line: err.line, column: err.column }
-            : null,
-          err.input.source
-        );
-        callback(wrappedError);
-      } else {
-        callback(err);
-      }
+    .catch(function(error) {
+      const preparedError =
+        error.name === "CssSyntaxError"
+          ? new CssLoaderError(
+              "Syntax Error",
+              error.reason,
+              error.line != null && error.column != null
+                ? { line: error.line, column: error.column }
+                : null,
+              error.input.source
+            )
+          : error;
+
+      callback(preparedError);
     });
 };
-
-function formatMessage(message, loc, source) {
-  var formatted = message;
-  if (loc) {
-    formatted = formatted + " (" + loc.line + ":" + loc.column + ")";
-  }
-  if (loc && source) {
-    formatted =
-      formatted + "\n\n" + formatCodeFrame(source, loc.line, loc.column) + "\n";
-  }
-  return formatted;
-}
-
-function CSSLoaderError(name, message, loc, source, error) {
-  Error.call(this);
-  Error.captureStackTrace(this, CSSLoaderError);
-  this.name = name;
-  this.error = error;
-  this.message = formatMessage(message, loc, source);
-  this.hideStack = true;
-}
-
-CSSLoaderError.prototype = Object.create(Error.prototype);
-CSSLoaderError.prototype.constructor = CSSLoaderError;
